Responsive
The top navigation didn't shrink down to fit smaller screen sizes.
So I started to problem solve by creating another css and html file. I learned that the bar had too much in it in terms  of content. Next step is make changes in order to display everything I need on the site.